+++ src/tags.cpp
@@ -2906,13 +2906,13 @@ namespace Exiv2 {
 
     float fnumber(float apertureValue)
     {
-        return static_cast<float>(std::exp(std::log(2.0) * apertureValue / 2));
+        return std::exp(std::log(2.0f) * apertureValue / 2.f);
     }
 
     URational exposureTime(float shutterSpeedValue)
     {
         URational ur(1, 1);
-        double tmp = std::exp(std::log(2.0) * shutterSpeedValue);
+        double tmp = std::exp(std::log(2.0) * static_cast<double>(shutterSpeedValue));
         if (tmp > 1) {
             ur.second = static_cast<long>(tmp + 0.5);
         }
